requirements.txt
python-binance==1.0.16
python-telegram-bot==13.15
PyYAML==6.0
pandas==1.5.3
numpy==1.24.3
requests==2.31.0
scikit-learn==1.2.2
streamlit==1.26.0
sqlalchemy==1.4.52
ta==0.11.2
matplotlib==3.7.1
